// src/enums/nft-gen-configurations.ts
export enum enumNFTGenConfig {
  NAME = "name",
  SUPPLY = "supply",
  FAMILY = "family",
  SYMBOL = "symbol",
  BASE_URL = "base_url",
  DESCRIPTION = "description",
  RENDER_FORMAT = "render_engine",
  OUTPUT_IMAGE_TYPE = "output_image_type",
  RENDER_FORMAT_WIDTH = "render_format_width",
  RENDER_FORMAT_HEIGHT = "render_format_height",
  RENDER_QUALITY = "render_quality",
  BLOCKCHAIN = "blockchain",
}